add user&login

+++ app/Http/Controllers/TasksController.php
@@ -47,12 +47,15 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content'=>'required|max:255',
             ]);
+            
+        
         $task=new Task;
         $task->status=$request->status;
         $task->content=$request->content;
+        $task->user_id=$request->user()->id;
         $task->save();
         
-        return redirect('/');
+        return redirect('tasks');
     }
 
     /**
@@ -64,9 +67,12 @@ class TasksController extends Controller
     public function show($id)
     {
         $task=Task::findOrFail($id);
-        
-        return view('tasks.show', ['task'=>$task,
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', ['task'=>$task,
         ]);
+        }
+        
+        return redirect('/');
     }
 
     /**
@@ -78,9 +84,11 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task=Task::findOrFail($id);
-        
-        return view('tasks.edit', ['task'=>$task,
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', ['task'=>$task,
         ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -98,11 +106,13 @@ class TasksController extends Controller
         'status' => 'required|max:10',
         'content'=>'required|max:255',
         ]);
-        $task=Task::findOrFail($id);
         
-        $task->status=$request->status;
-        $task->content=$request->content;
-        $task->save();
+        $task=Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            $task->status=$request->status;
+            $task->content=$request->content;
+            $task->save();
+        }
         
         return redirect('/');
     }
@@ -117,7 +127,11 @@ class TasksController extends Controller
     {
         $task=Task::findOrFail($id);
         
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('tasks');
+        }
+        
         return redirect('/');
     }
 }
